+++ api/src/com/cloud/network/tls/CertService.java
@@ -14,7 +14,7 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-package com.cloud.network.ssl;
+package com.cloud.network.tls;
 
 import java.security.cert.Certificate;
 import java.util.List;
